requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
openai==1.98.0
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.5.1 